+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow prop `ref` to be passed to `PhoneField`.
 
+### Fixed
+
+- Phone number not rendered when rule don't have a mask.
+
 ## [0.1.0] - 2020-03-19
 
 ### Added